Rename ModuleView to RecommendationsPreview
