+++ projects/plugins/boost/app/assets/src/js/modules/image-size-analysis/RecommendationsPreview.svelte
